Сделал оркестратор, берет из трех наших файлов и накладываетна одну html
