+++ .gitignore
@@ -6,3 +6,4 @@ __pycache__/
 
 Data/*
 !Data/.gitkeep
+/venv/
